bootcamp onwership

+++ controllers/bootcamps.js
@@ -29,34 +29,45 @@ exports.getBootcamp = asyncHandler(async (req, res, next) => {
 // @route /api/v1/bootcamps
 // @access private
 exports.createBootcamps = asyncHandler(async (req, res, next) => {
-  try {
-    const bootcamp = await Bootcamps.create(req.body);
-    res.status(201).json({
-      success: true,
-      data: bootcamp,
-    });
-  } catch (err) {
-    next(err);
+  req.body.user = req.user.id;
+
+  const publisherBootcamp = await Bootcamps.findOne({ user: req.user.id });
+  if (publisherBootcamp && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(
+        `The user with id ${req.user.id} have already publish the bootcamp`
+      )
+    );
   }
+  const bootcamp = await Bootcamps.create(req.body);
+  res.status(201).json({
+    success: true,
+    data: bootcamp,
+  });
 });
 // @desc update bootcamps
 // @route /api/v1/bootcamps/:id
 // @access public
 exports.updateBootcamps = asyncHandler(async (req, res, next) => {
   try {
-    const bootcamp = await Bootcamps.findByIdAndUpdate(
-      req.params.id,
-      req.body,
-      {
-        new: true,
-        runValidators: true,
-      }
-    );
+    let bootcamp = await Bootcamps.findById(req.params.id);
     if (!bootcamp) {
       return next(
         new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404)
       );
     }
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
+      return next(
+        new ErrorResponse(
+          `User ${req.params.id} is not authorize to update`,
+          401
+        )
+      );
+    }
+    bootcamp = await Bootcamps.findByIdAndUpdate(req.params.id, req.body, {
+      new: true,
+      runValidators: true,
+    });
     res.status(200).json({ success: true, data: bootcamp });
   } catch (err) {
     next(err);
@@ -70,6 +81,11 @@ exports.deleteBootcamps = asyncHandler(async (req, res, next) => {
   if (!bootcamp) {
     return next(
       new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404)
+    );
+  }
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(`User ${req.params.id} is not authorize to delete`, 401)
     );
   }
   await bootcamp.deleteOne();
@@ -111,7 +127,11 @@ exports.bootcampUploadPhoto = asyncHandler(async (req, res, next) => {
       new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404)
     );
   }
-
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(`User ${req.params.id} is not authorize to update`, 401)
+    );
+  }
   if (!req.files) {
     return next(new ErrorResponse(`Please upload a file`, 400));
   }
